closes #1805
